fix(userspace/libsinsp): Add a cast
Fix

userspace\libsinsp\event.h(456,1): warning C4267: 'argument': conversion from 'size_t' to 'uint32_t', possible loss of data [build\libsinsp\sinsp.vcxproj]

Visual C++ generates a bunch of similar warnings, but this minimizes a
lot of build noise.

+++ userspace/libsinsp/event.h
@@ -453,7 +453,7 @@ private:
 
 		for(j = 0; j < nparams; j++)
 		{
-			par.init((char*)params[j].buf, params[j].size);
+			par.init((char*)params[j].buf, (int)params[j].size);
 			m_params.push_back(par);
 		}
 	}
